+++ dio_cache_interceptor/lib/src/store/cache_store.dart
@@ -6,8 +6,25 @@ abstract class CacheStore {
   /// Checks if key exists in store
   Future<bool> exists(String key);
 
-  /// Retrieves cached response from store
+  /// Retrieves cached response from the given key.
   Future<CacheResponse?> get(String key);
+
+  /// Retrieves cached responses from a path pattern.
+  ///
+  /// [pathPattern] path pattern (e.g. RegExp('https://www.example.com/a/b') or
+  /// RegExp(r'https://www.example.com/a/\d+)).
+  ///
+  /// [queryParams] filter is processed in the following way:
+  /// - null: all entries are collected,
+  /// - null value: all entries containing the key are collected,
+  /// - otherwise key/value match only.
+  ///
+  /// You should be very restrictive when using this method as the underlying
+  /// store may parse and load all data from the store.
+  Future<List<CacheResponse>> getFromPath(
+    RegExp pathPattern, {
+    Map<String, String?>? queryParams,
+  });
 
   /// Pushes response in store
   Future<void> set(CacheResponse response);
@@ -16,6 +33,23 @@ abstract class CacheStore {
   /// [staleOnly] flag will remove it only if the key is expired
   /// (from maxStale).
   Future<void> delete(String key, {bool staleOnly = false});
+
+  /// Removes keys from the given filters.
+  ///
+  /// [pathPattern] path pattern (e.g. RegExp('https://www.example.com/a/b') or
+  /// RegExp(r'https://www.example.com/a/\d+)).
+  ///
+  /// [queryParams] filter is processed in the following way:
+  /// - null: all entries are collected,
+  /// - null value: all entries containing the key are collected,
+  /// - otherwise key/value match only.
+  ///
+  /// You should be very restrictive when using this method as the underlying
+  /// store may parse and load all data from the store.
+  Future<void> deleteFromPath(
+    RegExp pathPattern, {
+    Map<String, String?>? queryParams,
+  });
 
   /// Removes all keys from store.
   /// [priorityOrBelow] flag will remove keys only for the priority or below.
@@ -30,4 +64,29 @@ abstract class CacheStore {
 
   /// Releases underlying resources (if any)
   Future<void> close();
+
+  /// Checks if the given url matches with the given filters.
+  /// [url] must conform to uri parsing.
+  bool pathExists(
+    String url,
+    RegExp pathPattern, {
+    Map<String, String?>? queryParams,
+  }) {
+    if (!pathPattern.hasMatch(url)) return false;
+
+    var hasMatch = true;
+
+    final uri = Uri.parse(url);
+    if (queryParams != null) {
+      for (final entry in queryParams.entries) {
+        hasMatch &= (!uri.queryParameters.containsKey(entry.key));
+        if (entry.value != null) {
+          hasMatch &= !uri.queryParameters.containsValue(entry.value);
+        }
+        if (!hasMatch) break;
+      }
+    }
+
+    return hasMatch;
+  }
 }