+++ dio_cache_interceptor_db_store/lib/src/store/dio_cache_interceptor_db_store.dart
@@ -4,7 +4,7 @@ import 'package:dio_cache_interceptor_db_store/src/store/database.dart';
 /// A store saving responses in a dedicated database
 /// from an optional [directory].
 ///
-class DbCacheStore implements CacheStore {
+class DbCacheStore extends CacheStore {
   static const String tableName = 'dio_cache';
 
   /// Database name. Optional.
@@ -55,13 +55,43 @@ class DbCacheStore implements CacheStore {
   }
 
   @override
+  Future<void> deleteFromPath(
+    RegExp pathPattern, {
+    Map<String, String?>? queryParams,
+  }) async {
+    return _getFromPath(
+      pathPattern,
+      queryParams: queryParams,
+      onResponseMatch: (r) => delete(r.cacheKey),
+    );
+  }
+
+  @override
   Future<bool> exists(String key) {
     return _db.dioCacheDao.exists(key);
   }
 
   @override
-  Future<CacheResponse?> get(String key) async {
+  Future<CacheResponse?> get(String key) {
     return _db.dioCacheDao.get(key);
+  }
+
+  @override
+  Future<List<CacheResponse>> getFromPath(
+    RegExp pathPattern, {
+    Map<String, String?>? queryParams,
+  }) async {
+    final responses = <CacheResponse>[];
+
+    _getFromPath(
+      pathPattern,
+      queryParams: queryParams,
+      onResponseMatch: (r) async => responses.add(
+        _db.dioCacheDao.mapDataToResponse(r),
+      ),
+    );
+
+    return responses;
   }
 
   @override
@@ -72,5 +102,29 @@ class DbCacheStore implements CacheStore {
   @override
   Future<void> close() {
     return _db.close();
+  }
+
+  Future<void> _getFromPath(
+    RegExp pathPattern, {
+    Map<String, String?>? queryParams,
+    required Future<void> Function(DioCacheData) onResponseMatch,
+  }) async {
+    var results = <DioCacheData>[];
+    const limit = 10;
+    int? offset;
+
+    do {
+      final query = _db.dioCacheDao.select(_db.dioCacheDao.dioCache)
+        ..limit(limit, offset: offset);
+      results = await query.get();
+
+      for (final result in results) {
+        if (pathExists(result.url, pathPattern, queryParams: queryParams)) {
+          await onResponseMatch(result);
+        }
+      }
+
+      offset = (offset ?? 0) + limit;
+    } while (results.isNotEmpty);
   }
 }