+++ dio_cache_interceptor_db_store/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.2.0
+- feat: Add `Store.getFromPath` method.
+- feat: Add `Store.deleteFromPath` method.
+
 ## 4.1.1
 - core: Drift database re-generation to get rid of `Missing concrete implementation of 'getter ResultSetImplementation.attachedDatabase'.` message on pub.dev.